Update to unicode 16 (#41)
switch to u32 for LEXICON_OFFSETS, as unicode 16 makes
these bigger than can fit in a u16

+++ generator/src/lib.rs
@@ -341,7 +341,7 @@ fn write_codepoint_maps(ctxt: &mut Context, codepoint_names: Vec<(char, &str)>) 
         longest_name
     );
     ctxt.write_plain_string("LEXICON", &lexicon_string);
-    ctxt.write_debugs("LEXICON_OFFSETS", "u16", &lexicon_offsets);
+    ctxt.write_debugs("LEXICON_OFFSETS", "u32", &lexicon_offsets);
     ctxt.write_debugs("LEXICON_SHORT_LENGTHS", "u8", &lexicon_short_lengths);
     ctxt.write_debugs(
         "LEXICON_ORDERED_LENGTHS",
